pyproject.toml
[project]
name = "vr-datalake-user"
version = "0.1.0"
description = "Cloud Run Job for daily user image anonymization using RetinaFace and YOLO license plate detection"
authors = [
    {name = "mgonzalz",email = "maria.gonzalez@visualroad.net"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "google-cloud-bigquery>=3.20.0,<4.0.0",
    "google-cloud-storage>=2.16.0,<3.0.0",
    "huggingface-hub>=0.24.0,<1.0.0",
    "numpy>=1.26.0,<2.0.0",
    "opencv-python-headless>=4.9.0,<5.0.0",
    "pillow>=10.0.0,<11.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "retina-face>=0.0.17,<0.1.0",
    "tf-keras (>=2.15,<3.0)",
    "ultralytics>=8.3.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
]
notebook = [
    "jupyter>=1.0.0,<2.0.0",
    "matplotlib>=3.8.0,<4.0.0",
]

[tool.poetry]
packages = [
    {include = "anonymizer", from = "src"},
    {include = "gcs", from = "src"},
]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
